Unify TerminalLogger progress stripping code
